continuation RF home network

+++ DisplayUno2/DisplayUno2.ino
@@ -100,7 +100,7 @@ void loop() {
     loadsequencepicture();
   }
   
-  if (runningtiming - sensortiming > 15000){
+  if ((unsigned long)runningtiming - sensortiming > 15000){
     sensortiming = runningtiming;
 
     bool newval = read_sensors();
